wmviz/requirements.txt
streamlit>=1.32
plotly>=5.20
pandas>=2.1
databricks-sdk>=0.30
